added postprocessing for brightness

+++ src/debug_draw.cpp
@@ -4,7 +4,7 @@ void draw_debug_lines(std::vector<Vertex> line_vertices)
 
     D3D11_BUFFER_DESC vert_buffer_desc = {};
     vert_buffer_desc.Usage = D3D11_USAGE_DEFAULT;
-    vert_buffer_desc.ByteWidth = sizeof(Vertex)*line_vertices.size();
+    vert_buffer_desc.ByteWidth = (UINT)(sizeof(Vertex)*line_vertices.size());
     vert_buffer_desc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
     vert_buffer_desc.CPUAccessFlags = 0;
     vert_buffer_desc.MiscFlags = 0;
@@ -24,5 +24,5 @@ void draw_debug_lines(std::vector<Vertex> line_vertices)
 
     device_context->IASetInputLayout(vertex_layout); 
     device_context->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_LINELIST);
-    device_context->Draw(line_vertices.size(), 0);
+    device_context->Draw((UINT)line_vertices.size(), 0);
 }
